@@ -10,7 +10,7 @@ edit_on_github = True
 [options]
 zip_safe = False
 packages = find:
-python_requires = >=3.7
+python_requires = >=3.8
 setup_requires = setuptools_scm
 install_requires =
     numpy
@@ -23,11 +23,11 @@ install_requires =
     cobaya
     pyccl
     sacc
-    fgspectra @ git+https://github.com/simonsobs/fgspectra@act_sz_x_cib#egg=fgspectra
-    mflike @ git+https://github.com/simonsobs/lat_mflike@master
+    fgspectra>=1.1.0
+    syslibrary
 
 [options.package_data]
-soliket = *.yaml,*.bibtex,clusters/data/*,clusters/data/selFn_equD56/*,lensing/data/*.txt,mflike/*.yaml,tests/*.yaml,data/xcorr_simulated/*.txt,data/CosmoPower/CP_paper/CMB/*.pkl,szlike/*.yaml
+soliket = *.yaml,*.bibtex,clusters/data/*,clusters/data/selFn_equD56/*,lensing/data/*.txt,lensing/data/*.fits,mflike/*.yaml,tests/*.yaml,data/xcorr_simulated/*.txt,data/CosmoPower/CP_paper/CMB/*.pkl,szlike/*.yaml
 testpaths = "soliket"
 text_file_format = rst
 
@@ -39,7 +39,7 @@ docs =
 
 [flake8]
 select = E713,E704,E703,E714,E741,E10,E11,E20,E22,E23,E25,E27,E301,E302,E304,E9,
-		 F405,F406,F5,F6,F7,F8,W1,W2,W3,W6,E501
+		 F405,F406,F5,F6,F7,F8,E501
 max-line-length = 90
 exclude = .tox,build,cobaya_packages,test,.eggs
 
