@@ -1,21 +1,39 @@
-r"""Class to calculate bias models for haloes and galaxies as cobaya Theory classes
+"""
+.. module:: soliket.bias
 
+:Synopsis: Class to calculate bias models for haloes and galaxies as cobaya
+Theory classes.
+:author: Ian Harrison
+
+Usage
+-----
+
+To use the Linear Bias model, simply add it as a theory code alongside camb in
+your run settings, e.g.:
+
+.. code-block:: yaml
+
+  theory:
+    camb:
+    soliket.bias.linear_bias:
+
+
+Implementing your own bias model
+--------------------------------
+
+If you want to add your own bias model, you can do so by inheriting from the
+``soliket.Bias`` theory class and implementing your own custom ``calculate()``
+function (have a look at the linear bias model for ideas).
 """
 
-import pdb
 import numpy as np
-from typing import Sequence, Union
+from typing import Optional
 from cobaya.theory import Theory
-from cobaya.log import LoggedError
+from cobaya.typing import InfoDict
 
 
 class Bias(Theory):
-
-    kmax: float = 10.  # Maximum k (1/Mpc units) for Pk, or zero if not needed
-    nonlinear: bool = False  # whether to get non-linear Pk from CAMB/Class
-    z: Union[Sequence, np.ndarray] = []  # redshift sampling
-
-    extra_args: dict = {}  # extra (non-parameter) arguments passed to ccl.Cosmology()
+    """Parent class for bias models."""
 
     _logz = np.linspace(-3, np.log10(1100), 150)
     _default_z_sampling = 10**_logz
@@ -24,16 +42,15 @@ class Bias(Theory):
     def initialize(self):
 
         self._var_pairs = set()
-        # self._var_pairs = [('delta_tot', 'delta_tot')]
 
     def get_requirements(self):
         return {}
 
     def must_provide(self, **requirements):
 
-        options = requirements.get('linear_bias') or {}
+        options = requirements.get("linear_bias") or {}
 
-        self.kmax = max(self.kmax, options.get('kmax', self.kmax))
+        self.kmax = max(self.kmax, options.get("kmax", self.kmax))
         self.z = np.unique(np.concatenate(
                             (np.atleast_1d(options.get("z", self._default_z_sampling)),
                             np.atleast_1d(self.z))))
@@ -41,50 +58,46 @@ class Bias(Theory):
         # Dictionary of the things needed from CAMB/CLASS
         needs = {}
 
-        self.nonlinear = self.nonlinear or options.get('nonlinear', False)
+        self.nonlinear = self.nonlinear or options.get("nonlinear", False)
         self._var_pairs.update(
             set((x, y) for x, y in
-                options.get('vars_pairs', [('delta_tot', 'delta_tot')])))
+                options.get("vars_pairs", [("delta_tot", "delta_tot")])))
 
-        needs['Pk_grid'] = {
-                'vars_pairs': self._var_pairs or [('delta_tot', 'delta_tot')],
-                'nonlinear': (True, False) if self.nonlinear else False,
-                'z': self.z,
-                'k_max': self.kmax
+        needs["Pk_grid"] = {
+                "vars_pairs": self._var_pairs or [("delta_tot", "delta_tot")],
+                "nonlinear": (True, False) if self.nonlinear else False,
+                "z": self.z,
+                "k_max": self.kmax
             }
 
         assert len(self._var_pairs) < 2, "Bias doesn't support other Pk yet"
         return needs
 
     def _get_Pk_mm(self):
-
         for pair in self._var_pairs:
-
-            if self.nonlinear:
-                self.k, self.z, Pk_mm = self.provider.get_Pk_grid(var_pair=pair,
-                                                            nonlinear=True)
-                # Pk_mm = np.flip(Pk_nonlin, axis=0)
-            else:
-                self.k, self.z, Pk_mm = self.provider.get_Pk_grid(var_pair=pair,
-                                                         nonlinear=False)
-                # Pk_mm = np.flip(Pk_lin, axis=0)
+            self.k, self.z, Pk_mm = \
+                self.provider.get_Pk_grid(var_pair=pair, nonlinear=self.nonlinear)
 
         return Pk_mm
 
-    def get_Pk_gg_grid(self):
-        return self._current_state['Pk_gg_grid']
+    def get_Pk_gg_grid(self) -> dict:
+        return self._current_state["Pk_gg_grid"]
 
-    def get_Pk_gm_grid(self):
-        return self._current_state['Pk_gm_grid']
+    def get_Pk_gm_grid(self) -> dict:
+        return self._current_state["Pk_gm_grid"]
 
 
 class Linear_bias(Bias):
+    """
+    :Synopsis: Linear bias model.
+    
+    Has one free parameter, :math:`b_\mathrm{lin}` (``b_lin``).
+    """
 
-    params = {'b_lin': None}
-
-    def calculate(self, state, want_derived=True, **params_values_dict):
+    def calculate(self, state: dict, want_derived: bool = True,
+                  **params_values_dict) -> Optional[bool]:
 
         Pk_mm = self._get_Pk_mm()
 
-        state['Pk_gg_grid'] = params_values_dict['b_lin']**2. * Pk_mm
-        state['Pk_gm_grid'] = params_values_dict['b_lin'] * Pk_mm
+        state["Pk_gg_grid"] = params_values_dict["b_lin"]**2. * Pk_mm
+        state["Pk_gm_grid"] = params_values_dict["b_lin"] * Pk_mm
